dockerized

+++ src/main/java/ec/edu/espe/arquitectura/examen/dto/EmpleadosPagoRQ.java
@@ -1,6 +1,6 @@
 package ec.edu.espe.arquitectura.examen.dto;
 
-import java.math.BigDecimal;
+
 
 import lombok.Builder;
 import lombok.Data;
@@ -9,5 +9,5 @@ import lombok.Data;
 @Builder
 public class EmpleadosPagoRQ {
     private String numeroCuenta;
-    private BigDecimal valor;
+    private Double valor;
 }
