+++ src/main/java/ec/edu/espe/arquitectura/examen/model/EmpleadosPago.java
@@ -1,6 +1,5 @@
 package ec.edu.espe.arquitectura.examen.model;
 
-import java.math.BigDecimal;
 
 import lombok.Builder;
 import lombok.Data;
@@ -9,6 +8,6 @@ import lombok.Data;
 @Builder
 public class EmpleadosPago {
     private String numeroCuenta;
-    private BigDecimal valor;
+    private Double valor;
     private String estado;
 }
